+++ R/platjam-import.R
@@ -388,6 +388,11 @@ coverage_matrix2nmat <- function
 #'    `"abs_mean"` the mean of absolute values.
 #' @param ylims `vector` of maximum y-axis values for each heatmap profile;
 #'    or `list`
+#' @param border `logical` indicating whether to draw a border around the
+#'    heatmap, which includes all heatmap panels in the event of
+#'    splitting by clustering. The `border` can be supplied as a vector,
+#'    so the `border` can be applied specifically to each heatmap
+#'    if needed.
 #' @param iter.max integer value indicating the maximum iterations
 #'    performed by k-means clustering, only relevant when `k_clusters`
 #'    is non-zero.
@@ -460,6 +465,7 @@ nmatlist2heatmaps <- function
  ht_gap=grid::unit(7, "mm"),
  profile_value=c("mean", "sum", "abs_mean", "abs_sum"),
  ylims=NULL,
+ border=TRUE,
  iter.max=20,
  use_raster=TRUE,
  do_plot=TRUE,
@@ -477,6 +483,11 @@ nmatlist2heatmaps <- function
    if (length(main_heatmap) == 0 || main_heatmap > length(nmatlist)) {
       main_heatmap <- 1;
    }
+
+   if (length(border) == 0) {
+      border <- FALSE;
+   }
+   border <- rep(border, length.out=length(nmatlist));
 
    ## k_method
    kmeans <- stats::kmeans;
@@ -1036,10 +1047,7 @@ nmatlist2heatmaps <- function
          }
       }
       if (!is.function(color) && length(color) == 1 && divergent) {
-         hcl <- jamba::col2hcl(color);
-         hcl["H",] <- (180 + hcl["H",]) %% 360;
-         hcl["C",] <- jamba::noiseFloor(hcl["C",], minimum=160);
-         color2 <- jamba::hcl2col(hcl);
+         color2 <- color_complement(color, ...);
          color <- c(color2, "white", color);
          divergent <- TRUE;
       }
@@ -1114,6 +1122,7 @@ nmatlist2heatmaps <- function
          pos_line=FALSE,
          use_raster=use_raster,
          col=colramp,
+         border=border[[i]],
          top_annotation=ComplexHeatmap::HeatmapAnnotation(
             lines=EnrichedHeatmap::anno_enriched(
                gp=grid::gpar(col=k_colors),
@@ -1452,4 +1461,102 @@ get_nmat_ceiling <- function
          round(digits=3, iceiling));
    }
    return(iceiling);
+}
+
+#' Create color complement by rotating the color hue
+#'
+#' Create color complement by rotating the color hue
+#'
+#' This function rotates the color hue to create a complementary
+#' color for each `color` input. It differs from standard methods
+#' by using warped color hue by default (`useWarpHue=TRUE`), which
+#' uses a red-yellow-blue color wheel instead of R default
+#' red-green-blue. It also imposes a minimum chroma, which
+#' ensures the output color is reasonably high in color
+#' saturation.
+#'
+#' @family jam utility functions
+#'
+#' @param color `character` vector of R compatible colors.
+#' @param Hflip numeric value in degrees (from 0 to 360) added
+#'    to the color hue to produce the final color hue.
+#' @param Cfloor numeric value used to limit output chroma `C`
+#'    values to this minimum value.
+#' @param Lrange `numeric` vector with the allowed range of output
+#'    luminance `L` values. When supplied, output values are
+#'    simply forced to this range with no other scaling of intermediate
+#'    values.
+#' @param useWarpHue `logical` indicating whether to use the warp
+#'    hue functions `colorjam::h2hw()` and `colorjam::hw2h()` which
+#'    effectively change the color wheel from red-green-blue to
+#'    red-yellow-blue.
+#' @param ... additional arguments are ignored.
+#'
+#' @examples
+#' n <- 5;
+#' rc <- colorjam::rainbowJam(n);
+#' rc_comp <- color_complement(rc, Cfloor=180);
+#' rc_comp2 <- color_complement(rc, Cfloor=180, useWarpHue=FALSE);
+#' jamba::showColors(list(rainbowJam=rc,
+#'    `complement_warped\n(default uses RYB)`=rc_comp,
+#'    `complement_nowarp\n(unwarped uses RGB)`=rc_comp2));
+#'
+#' ## divergent color gradients through white
+#' ## hint: use higher lens value to make middle colors more intense
+#' rc_ramps <- lapply(jamba::nameVector(seq_along(rc)), function(i){
+#'    j <- jamba::getColorRamp(c(rc[i], "white", rc_comp[i]),
+#'       n=25,
+#'       lens=0,
+#'       divergent=TRUE);
+#'    names(j) <- "";
+#'    names(j)[1] <- "original colors";
+#'    names(j)[25] <- "color complements";
+#'    j;
+#' });
+#' jamba::showColors(rc_ramps, groupCellnotes=TRUE, groupByColors=FALSE);
+#'
+#' ## divergent color gradients through white
+#' ## hint: use higher lens value to make middle colors more intense
+#' rc_ramps2 <- lapply(jamba::nameVector(seq_along(rc)), function(i){
+#'    j <- jamba::getColorRamp(c(rc[i], "black", rc_comp[i]),
+#'       n=25,
+#'       lens=1,
+#'       divergent=TRUE);
+#'    names(j) <- "";
+#'    names(j)[1] <- "original colors";
+#'    names(j)[25] <- "color complements";
+#'    j;
+#' });
+#' jamba::showColors(rc_ramps2, groupCellnotes=TRUE, groupByColors=FALSE);
+#'
+#' @export
+color_complement <- function
+(color,
+ Hflip=180,
+ Cfloor=160,
+ Lrange=c(0, 100),
+ useWarpHue=TRUE,
+ ...)
+{
+   hcl <- jamba::col2hcl(color);
+   H <- hcl["H",];
+   if (useWarpHue) {
+      H <- colorjam::h2hw(H);
+   }
+   newH <- (Hflip + H) %% 360;
+   if (useWarpHue) {
+      newH <- colorjam::hw2h(newH);
+   }
+   hcl["H",] <- newH;
+   if (length(Cfloor) > 0) {
+      hcl["C",] <- jamba::noiseFloor(hcl["C",],
+         minimum=Cfloor);
+   }
+   if (length(Lrange) > 0) {
+      hcl["L",] <- jamba::noiseFloor(hcl["L",],
+         minimum=min(Lrange),
+         ceiling=max(Lrange));
+   }
+   color2 <- jamba::hcl2col(hcl);
+   return(color2);
 }